refactor: 패키지 구조 DDD에 맞춰 리팩토링
