+++ com.delivery-signal.eureka.client.order/src/main/java/com/delivery_signal/eureka/client/order/application/dto/response/OrderForDeliveryResponseDto.java
@@ -1,4 +1,4 @@
-package com.delivery_signal.eureka.client.order.presentation.internal.dto.response;
+package com.delivery_signal.eureka.client.order.application.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
